Update Calendar-Script.ps1
Adding logging functionality to the script and allow it to be invoked before running the initial main menu.

Currently none of the functions are implementing the logging functionality but the functionality has been built for this commit.

For more information regarding how logging is being done please check the wiki article regarding logging